PATCH: [ 3088720 ] Add OpenCandy to the InnoSetup installer
This is using the OpenCandy 1.4.2.28 Inno SDK.

The WinMerge.iss is including a file OpenCandyKeys.iss which is not included. This file contains the private keys for WinMerge, and can be obtained from the WinMerge administrators if you are creating the WinMerge releases.

If you just want to use the installer without the WinMerge keys, then remove this line:
#include <OpenCandyKeys.iss>
This makes the installer compile with default demo keys instead of the WinMerge specific keys.

--HG--
branch : R2_14