Improve reading of files > 2GB
* Now that commit #4e4e46c has allowed files larger than 2GB, there are
number of places in WinMerge that are not prepared to handle files that
large.  These will have to be discovered by "trial and error".  These
are the first two problems encountered ...

* This patch repairs two places that `assert`with long file lengths.
* It should be noted that Windows can only read/write in chunks of up to
2GB.  The code around lines 227 to 232 is part of a loop that is
capable of reading very large files, and now has the correct setup code
to limit the size of each successive read.
* The `assert` at line 911 is completely unnecessary for both 32- and
64-bit builds.

--HG--
branch : jp