Improve handling of last lines in files (5)
- When using Copy Left (or Copy Right) on a block of line differences at
the _very_ end of both files, a fatal abort would occur when copying
lines into the shortest of the two end blocks.

- This fatal error would occur when deleting "ghost" lines at the end
of that shortest block, because the starting location of the "ghost"
lines was improperly calculated.

- There remains a non-fatal bug in that an extra EOL (and thus an
extra blank line) is added to the end of a Copy (Left or Right) when
that action is on the text block at the very end of the file.

- The Undo action (ctrl-Z) continues to function correctly for these
situations.