Attempt to fix the issue #100 - Part 2
- Commit
[2f8e009](https://bitbucket.org/winmerge/winmerge/commits/2f8e009c3f473ca0f6a639990d6307ed95d7694e)
checks for the existence of the `WinIMergeLib.dll` and abandons the
**ImageCompareTest.Open** test if the `.dll` cannot be loaded
correctly.  This checking prevents the inappropriate (and seriously
misleading) SEH error later in the code; _**however**_ it allows the test to
Pass even when the `.dll` is not present.

This follow-on patch forces the test to Fail cleanly, with a
relevant message, when the `.dll` cannot be loaded correctly.

- Also, the handling of `EXPECT_NE(nullptr, pDoc)` in other tests is
improved to prevent SEH errors if `pDoc==nullptr` should ever become
true.