Reduce Level 4 Build Warnings (1)
I'm starting on a quest to eliminate as many Level=4 build warnings as
makes sense (and is reasonably easy).  I will not actually "release"
the patches to the `*.vcxproj`files that change the warning level of
the builds (which remain at Level 3); maybe someday ...

I'm not addressing changes to any of the `poco\*` software, except
perhaps a few of the `*.h` files that generate warnings when included
into a WinMerge source file.

* Warning C4100 'a': unreferenced formal parameter
* Warning C4189 'b': local variable is initialized but not referenced

These warnings are suppressed simply by disabling them in the
appropriate `*.vcxproj` file.  I disable them because I claim that
(because of conditional compilation, etc) they are normal and
reasonable coding practice.