Last Line of files vs "Ghost Lines" (1)
The special case of inserting text into the very last line of a file
when that last line is marked as LF_GHOST.  Effectively, the new text
is supposed to go "before" the Ghost, but mechanically the text is
inserted into the Ghost itself, with a new Ghost line appearing at the
end of the file.

Later, the Ghost status of both the first and last inserted lines will
get straightened out, with the trailing Ghost line becoming a NULL
line.  By setting the last line to NULL, the line will eventually be
removed or become an actual LF_GHOST line in
`CMergeDoc::PrimeTextBuffers()`