Last Line of Files vs "Ghost Lines" (2)
Handle the special case of `Undo()` at the very end of a file.

`Undo()` checks that the text that is being replaced in the file
matches the records being kept in the `m_aUndoBuf[]` structure.  The
deletion of text at very end of the file necessary involves removing an
additional CRLF (because of an LF_GHOST line which was removed too)
which causes this Undo comparison-checking to fail.

Following failure, and only if the failure is at the very end of the
file, a new secondary check is made to check for the deletion of the
additional CRLF.

I don't particularly like this code because it puts knowledge of
"Ghost" lines directly into the code of the `CCrystalTextBuffer` class.
A subsequent patch will factor this into virtual procedures shared with
the `CGhostTextBuffer` class.