Last Line of Files vs "Ghost Lines" (4)
Re-implements the Undo handling for the last line (see my final comments
in the earlier "...Ghost Lines (2)" commit) to move the "Ghost"
sensitive parts out of the `CCrystalTextBuffer` class and into the
`GhostTextBuffer` class.

A new virtual procedure `UndoInsert()`now exists in both
`CCrystalTextBuffer` and `GhostTextBuffer`, with the WinMerge and Ghost
specific checking in the `GhostTextBuffer`class, and the basic deletion
code in the `CCrystalTextBuffer`class.

Also, modify the `VERIFY()`testing in the Undo paths to correctly
interface with the general failure mechanism within `Undo()`