Adapt to latest VS2017 version 15.8
* As is often the case, the boost library needs to update its warning
message for the new compiler version.

* There is a bug in the gtest software; it is using a non-portable
`operator<<` overload.  The ambiguity of this usage is now being
flagged as an error by version 15.8.  There have been many updates to
the gtest software since the version incorporated into WinMerge, but
fortunately the "accepted" and merged solution to this problem simply
drops into WinMerge without issue (although I've added some comments
referencing the gtest PR#1620 -
https://github.com/google/googletest/pull/1620)

* Unfortunately, in order to trigger the source code in gtest PR#1620,
the value of the predefined macro `__cplusplus` must be accurate.  For
various reasons (see
https://blogs.msdn.microsoft.com/vcblog/2018/04/09/msvc-now-correctly-reports-__cplusplus/)
the compiler option `Zc:__cplusplus` must be set in order to generate
correct values for `__cplusplus`.  The changes to the
`*.vs2017.vcxproj`file accomplish this.  Note that the elimination of
the "additional option" `/EHa` is acceptable because that option is
properly specified elsewhere in the project's build properties.