"Private Build" is now easy and useful for developers
* A Private Build string can now be #defined in `version.h`; the
default value is an empty string (as it should be for all released
versions of **WinMerge**).   This value propagates into the
VersionInfo block via the `Merge2.rc` file, and is therefore available
to all checking of the VersionInfo block of the `.exe` file.

* The About dialog is modified to remove the separate line that
displayed the "Private Build:" information in the past, as that line
overlapped the "Copyright" text immediately below.  The "Private Build"
info is now displayed on the same line as the Version, separated by
" + ".  There is room for about 11 characters of "Private Build" text.

* The Configuration Log File now indicates the Private Build
information following the Version, rather that simply stating
"(Private Build)"

* Cleanup the mechanics of getting the "Private Build" information into
the About dialog.

* Also a few trivial code and comment changes.