"Open Dialog Auto-Completion" == "From MRU list" now fixed
 - The option "Open Dialog Auto-Completion" (on the "Options
(General)" setting page) failed to work correctly when set to the value
"From MRU list".  This appears to have been true since the
`2.13.11+-jp-8` commit on 04/18/2010 when the ComboBoxes for
file/folder selection were changed to use the the "ComboBoxEx32" common
control.  The failure occurs because `FindString()`, which is used in
`CSuperComboBox::OnEditchange()`, is only valid for CComboBox but not
CComboBoxEx.

 - This patch restores the "From MRU list" functionality, but
necessarily builds it upon the recent "Improvements to CSuperComboBox"
changes.

- It should also be noted that "From MRU list" provides a very limited
functionality when compared to the "From file system".  This was
true even with WinMerge versions 2.13.10 and earlier.