Remove `nFinalNullLines` checking, disabled by an earlier commit
* Github Commit 1b52d84 "MergeDoc.cpp: Fix github issue #106 ..." effectively removed this code
with a `#if 0`.  This present commit actually removes the code, but preserves ASSERT checking to
make sure that all `m_ptBuf[]` structures have the exact same length.

* These lines of code (from GitHub Commit 51519fc, 16 Jun 2018) were an early attempt to solve
various problems with the last line(s) of a comparison.  Subsequent changes made these lines
irrelevant.  Independently they were shown to be erroneous.