Fix obscure memory leak
A memory leak can be caused under the following circumstance:

 * Compare a directory structure with a number of folders and/or files.
 * Select one (or more) items.
 * Outside of WinMerge (i.e. without WinMerge's knowledge) delete one (or more) of the files or
folders that is already selected in WinMerge.
 * Switch back to WinMerge, and then select "View | Refresh Selected" from menu (or type
`Ctrl-F5`) to rescan only the selected items (note: some of these are now missing files or folders)

WinMerge will correctly detect that the files/folders are missing while rescanning, and will
modify its internal directory structure tree accordingly.  HOWEVER - WinMerge will **not**
actually `delete` the disconnected tree item's memory.  Thus causing a memory leak.  This will be
visible with a Debug build of WinMerge.

The fix is both simple and allows the surrounding code to be cleaned up a little bit.