Fix bitbucket issue #162: Regression: Incorrectly shows 'no changes' for large portions of certain binary files!
The bug of issue # 112 was introduced in commit d1a48cb and fixed in this commit and commit 2f202d1.

The cause of this problem is that diffutils recognizes the binary file submitted in issue # 112 as a text file of UCS4LE with BOM.
The merge editor temporarily converts to a UCS2LE file with BOM to compare the loaded files. However, diffutils incorrectly identified it as UCS4LE with BOM because there were 4 NUL characters at the beginning of the file.

There is confusion because diffutils sees it as UCS4LE and the merge editor sees it as UCS2LE.

Also, the current modified diffutils now give up on binary file comparisons.

In order to avoid the above two, convert the file to UTF-8 with BOM so that diffutils considers it to be a UTF-8 text file.