Improve vertical alignment string (#1200)
I experienced the same issue as in: https://github.com/WinMerge/winmerge/issues/1008

Both that user and I were misled by the "Match similar lines" string.  In English, it's confusing within the context of a comparison application because "match" is also a comparison term.

As such, I improved it to be "Align similar items".