Create Midnight.ini (#1430)
* Create Dark Mode.ini

Added an additional color scheme that is easy on the eyes.

* Update and rename Dark Mode.ini to Midnight.ini

 Changed The colors of the folder window to better differentiate this scheme from Solarized Dark.

* Update Midnight.ini

I've altered the value of the background for diff items in the folder window to better contrast with the text. I didn't want increase the contrast too high here, as that would defeat the purpose of a dark theme, but I think this is a good compromise between ease of readability and lack of eye strain.