Syntax/Pascal: Check for text behind multi-line string start
The start of a multi-line string is three single quotation marks (''') and then a newline (however spaces and tabs are ok). So check if that is the case.
Otherwise ``Test := '''a';`` would be treated as the start of a multi-line string instead of a single quotation mark and an a.