Syntax/Pascal: Support multiline strings (#2731)
* Syntax/Pascal: Support multiline strings

Delphi introduced multiline strings in version 12. They start with ''' and end with '''.

* Syntax/Pascal: Interpret '''' correctly

The string '''' was not interpreted as a single quotation mark but instead as the start of a multiline string.

* Syntax/Pascal: Correctly interpret Text inside multiline string

Inside a multiline string there is no other syntax highlighting and a single quotation mark (') is valid and does not end the string. The multiline string only ends if a corresponding end (''') is found.

* Syntax/Pascal: Check for text behind multi-line string start

The start of a multi-line string is three single quotation marks (''') and then a newline (however spaces and tabs are ok). So check if that is the case.
Otherwise ``Test := '''a';`` would be treated as the start of a multi-line string instead of a single quotation mark and an a.

* Syntax/Pascal: Check for text before multi-line string end

There may not be any character in front of the end of the multi-line string. This ensures, that the multi-line coloring is not disrupted with single quotation marks.